requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
backports.tarfile==1.2.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
gnupg==2.3.1
greenlet==3.2.3
importlib_metadata==8.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jeepney==0.9.0
Jinja2==3.1.6
MarkupSafe==3.0.2
more-itertools==10.7.0
pillow==11.3.0
psutil==7.0.0
pycparser==2.22
PyNaCl==1.5.0
pyotp==2.9.0
pysqlcipher3==1.2.0
qrcode==8.2
SecretStorage==3.3.3
SQLAlchemy==2.0.41
sqlcipher3==0.5.4
typing_extensions==4.14.1
uuid==1.30
Werkzeug==3.1.3
zipp==3.23.0
